OpenMP to configure pre-draw
Profiling on a DELL WYSE 3040 (Intel Atom x5-Z8350) showed significant CPU bottle neck. Partially immediate mode OpenGL and partially the prep for the draw.

Restricting to OpenGL 1.1 is desired for maximum compatibility and batching hindered performance so is not included. OpenMP for parallelising pre-draw setup nearly doubled the worst case FPS on the WYSE.